fix uninitialized data access in update_ptrs
it was possible for an object to be attempted allocated, for the
allocation to then cause a compaction, and then the object could be in
an uninitialized state while the compaction/update_ptrs step happened.
Fixed by unexporting the Arena::alloc function and using Arena::put
instead, which guarantees that objects will be initialized during a
potential update_ptrs fase.